+++ src/main/java/com/tcommerce/TCommerce/interfaces/dto/commerce/product/UpdateProductRequest.java
@@ -3,7 +3,6 @@ package com.tcommerce.TCommerce.interfaces.dto.commerce.product;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.Size;
 import org.springframework.web.multipart.MultipartFile;
-import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.List;
 
@@ -14,7 +13,7 @@ public record UpdateProductRequest(
         String description,
 
         @Min(value = 0, message = "Price must be greater than or equal to 0")
-        BigDecimal price,
+        BigInteger price,
 
         String categoryId,
