+++ src/main/java/com/tcommerce/TCommerce/interfaces/dto/sales/CartResponse.java
@@ -1,6 +1,6 @@
 package com.tcommerce.TCommerce.interfaces.dto.sales;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -8,7 +8,7 @@ public record CartResponse(
     String id,
     String userId,
     List<CartItemResponse> items,
-    BigDecimal totalPrice,
+    BigInteger totalPrice,
     LocalDateTime createdAt,
     LocalDateTime updatedAt
 ) {}
