+++ src/main/java/com/tcommerce/TCommerce/interfaces/dto/sales/OrderResponse.java
@@ -2,7 +2,7 @@ package com.tcommerce.TCommerce.interfaces.dto.sales;
 
 import com.tcommerce.TCommerce.domain.entities.sales.OrderStatus;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -11,7 +11,7 @@ public record OrderResponse(
     String userId,
     OrderStatus status,
     List<OrderItemResponse> items,
-    BigDecimal totalAmount,
+    BigInteger totalAmount,
     LocalDateTime createdAt,
     LocalDateTime updatedAt
 ) {}
